Refactor authentication domain by removing legacy files and consolidating auth utilities
- Remove entire auth-legacy directory and migrate remaining utilities to core/domains/auth
- Update import paths across multiple files to use new auth domain location
- Add GenerateJwtSecret command for JWT secret management
- Consolidate authentication exceptions and utility functions
- Clean up and reorganize authentication-related code structure
