--- src/core/domains/auth-legacy/consts/authConsts.ts
@@ -1,38 +1,0 @@
-/**
- * Constants for auth routes
- */
-const authConsts = {
-
-    /**
-     * Route for creating a new user
-     */
-    routes: {
-
-        /**
-         * Route for creating a new user
-         */
-        authCreate: 'authCreate',
-
-        /**
-         * Route for logging in
-         */
-        authLogin: 'authLogin',
-
-        /**
-         * Route for retrieving the current user
-         */
-        authUser: 'authUser',
-
-        /**
-         * Route for revoking a token
-         */
-        authRevoke: 'authRevoke',
-
-        /**
-         * Route for updating the current user
-         */
-        authUpdate: 'authUpdate'
-    }
-}
-
-export default authConsts
