--- src/core/domains/auth-legacy/enums/RolesEnum.ts
@@ -1,6 +1,0 @@
-const ROLES = Object.freeze({
-    USER: 'user',
-    ADMIN: 'admin'
-})
-
-export default ROLES
