--- src/core/domains/auth-legacy/utils/decodeJwt.ts
@@ -1,14 +1,0 @@
-import jwt from 'jsonwebtoken'
-
-import { IJSonWebToken } from '../../auth/interfaces/jwt/IJsonWebToken'
-
-/**
- * Decodes a JWT token using the provided secret.
- *
- * @param {string} secret - The secret to use to decode the token.
- * @param {string} token - The JWT token to decode.
- * @returns {IJSonWebToken} The decoded token.
- */
-export default (secret: string, token: string): IJSonWebToken => {
-    return jwt.verify(token, secret) as IJSonWebToken
-}
