+++ src/core/domains/auth/commands/GenerateJwtSecret.ts
