+++ src/core/domains/auth/exceptions/ForbiddenResourceError.ts
