+++ src/core/domains/auth/exceptions/InvalidJWTSecret.ts
