+++ src/core/domains/auth/exceptions/RateLimitedExceededError.ts
