+++ src/core/domains/auth/exceptions/UnauthorizedError.ts
