+++ src/core/domains/auth/providers/AuthProvider.ts
@@ -3,7 +3,9 @@
 import { aclConfig } from "@src/config/acl";
 import { authConfig } from "@src/config/auth";
 import BaseProvider from "@src/core/base/Provider";
+import { app } from "@src/core/services/App";
 
+import GenerateJwtSecret from "../commands/GenerateJwtSecret";
 import Auth from "../services/AuthService";
 
 class AuthProvider extends BaseProvider{
@@ -12,12 +14,17 @@ class AuthProvider extends BaseProvider{
         const authService = new Auth(authConfig, aclConfig);
         await authService.boot();
         
+        // Bind services
         this.bind('auth', authService);
         this.bind('auth.jwt', (() => authService.getDefaultAdapter())())
         this.bind('auth.acl', (() => authService.acl())())
+
+        // Register commands
+        app('console').register().register(GenerateJwtSecret)
     }
 
 
 }
 
 export default AuthProvider;
+
